☀️ 20% – Recover callee return types from the bytes: 7 more owners
THE CORRECTION. The previous commit read the rule one way round: a caller that
discards a result wants `void`. That is half of it. `resource_376:02000190`
closes by making its callee `s32` -- the OPPOSITE edit, on a call whose result
is equally discarded. Both directions occur, which means the return type is not
inferable from the call site at all.

It is, however, decidable from the reference bytes. A non-void return keeps a
value live across the argument setup and reverses two pre-call moves, so the
emitted order reports the type. `retype` flips one declaration at a time and
keeps what pays: 7 owners closed, 121 improved, converging in two passes.

WHY THIS IS NOT A SEARCH. Every declaration is scored against known bytes, the
candidate set per declaration is seven types, and a file has a handful of
declarations. It is a decision procedure over the interface, not a walk over
source shapes -- which is why it converges instead of plateauing.

Type widths went the same way, from the ledger's own rule that narrow types are
usually the wrong reading: widening narrow locals and narrow callee parameters
improved 37 rows. None closed, so no bytes yet, but they are structurally
correct and leave the residuals smaller for whatever closes them.

266,536 -> 267,272 bytes, 19.79% -> 19.84%. `make verify` green.

+++ semantic/resource_37b_c_02002244.c
@@ -3,18 +3,18 @@
 
 void Func_0200469e(void);
 void Func_0200469e_b(s32, s32, s32, s32, s32, s32);
-s32 Func_02004684(s32);
-s32 Func_0200468e(s32);
+s32 Func_02004684(u8);
+s32 Func_0200468e(u8);
 void Func_0200477e(s32, s32);
 void Func_02004798(s32, s32, s32, s32);
 void Func_020047a4(void);
 void Func_020047ea(s32);
-s32 Func_020046d4(s32);
+s32 Func_020046d4(u8);
 void Func_020046ba(s32, s32, s32, s32, s32, s32);
 void Func_020047b4(s32, s32, s32);
 void Func_02004712(s32);
 void Func_02004708(s32);
-s32 Func_02004706(s32);
+s32 Func_02004706(u8);
 void Func_02002372(void);
 void Func_0200473a(void);
